Show or hide leaflet aoi layer

Function to toggle the aoi layer on an existing leaflet map.

ff_layer_aoi(m, show = TRUE)

Arguments

m

An initialized leaflet map object or leafletProxy object.

show

A boolean value indicating whether the function call will be adding the layer to the map (TRUE) or removing the layer from the map (FALSE). Designed to be changed via shiny checkbox input by calling the function inside an observer.

Examples

# show the layer on a leaflet map object ("m")
m <- ff_make_leaflet()
m |> ff_layer_aoi(show = TRUE)

# hide the layer on a leaflet map object ("m")
m |> ff_layer_aoi(show = FALSE)

# use as part of a Shiny app with map "mainMap" and a boolean selector "show_aoi"
if(FALSE){
  shiny::observe({
    proxy <- leaflet::leafletProxy("mainMap")
    proxy |> ff_layer_aoi(show = input$show_aoi)
  })
}